ButcherStation: compatibility with the new beta version of the game, some fields have been removed and added.
removed a patch to fix the fish delivery point.
removed a patch to fix the bugged animation of the station when the rancher prematurely leaves, for example, to eat.
because the developers themselves fixed these errors.
the drop of additional meat has been rewritten, the drop list is checked every time, for better compatibility with the late added drop and other mods. crab meat has been added.