VaricolouredBalloons: Completely rewritten for the new skin system.
Now the mod does a complete randomization of the ball skin, choosing from all available (unlocked) skins if the Duplicant Artist has the assignment of a default red skin.
Added a Puft mutants skin - based on the appearance of the first version of the mod.
Added an Orange skin - based on the appearance of the Klei new skins.