BetterPlantTending: Unlocked Bonbon Tree mutations.
Mutated seeds are obtained when Seal drink Nectar from an non-mutated Bonbon Tree.
Now mutations affect the Nectar production rate: the higher the Yield Amount and the shorter the Life Cycle, the higher the production rate.
Fixed Duplicants could not manually extract Nectar from a Tree planted in a Planter Box.
Now the mutated Arbor Tree does not produce seeds.
Some Options are not displayed if the user does not own the corresponding DLC.